fix: change goto declarations provider to references provider
