+++ crates/tinymist-query/src/analysis/matcher.rs
@@ -1,0 +1,69 @@
+use log::debug;
+use typst::syntax::{
+    ast::{self, AstNode},
+    LinkedNode, SyntaxKind,
+};
+
+pub fn deref_lvalue(mut node: LinkedNode) -> Option<LinkedNode> {
+    while let Some(e) = node.cast::<ast::Parenthesized>() {
+        node = node.find(e.expr().span())?;
+    }
+    Some(node)
+}
+
+pub enum DerefTarget<'a> {
+    VarAccess(LinkedNode<'a>),
+    Callee(LinkedNode<'a>),
+    ImportPath(LinkedNode<'a>),
+}
+
+impl<'a> DerefTarget<'a> {
+    pub fn node(&self) -> &LinkedNode {
+        match self {
+            DerefTarget::VarAccess(node) => node,
+            DerefTarget::Callee(node) => node,
+            DerefTarget::ImportPath(node) => node,
+        }
+    }
+}
+
+pub fn get_deref_target(node: LinkedNode) -> Option<DerefTarget> {
+    let mut ancestor = node;
+    while !ancestor.is::<ast::Expr>() {
+        ancestor = ancestor.parent()?.clone();
+    }
+    debug!("deref expr: {ancestor:?}");
+    let ancestor = deref_lvalue(ancestor)?;
+    debug!("deref lvalue: {ancestor:?}");
+
+    let may_ident = ancestor.cast::<ast::Expr>()?;
+    if !may_ident.hash() && !matches!(may_ident, ast::Expr::MathIdent(_)) {
+        return None;
+    }
+
+    Some(match may_ident {
+        // todo: label, reference
+        // todo: import
+        // todo: include
+        ast::Expr::FuncCall(call) => DerefTarget::Callee(ancestor.find(call.callee().span())?),
+        ast::Expr::Set(set) => DerefTarget::Callee(ancestor.find(set.target().span())?),
+        ast::Expr::Ident(..) | ast::Expr::MathIdent(..) | ast::Expr::FieldAccess(..) => {
+            DerefTarget::VarAccess(ancestor.find(may_ident.span())?)
+        }
+        ast::Expr::Str(..) => {
+            let parent = ancestor.parent()?;
+            if parent.kind() != SyntaxKind::ModuleImport {
+                return None;
+            }
+
+            return Some(DerefTarget::ImportPath(ancestor.find(may_ident.span())?));
+        }
+        ast::Expr::Import(..) => {
+            return None;
+        }
+        _ => {
+            debug!("unsupported kind {kind:?}", kind = ancestor.kind());
+            return None;
+        }
+    })
+}
