+++ crates/tinymist/src/init.rs
@@ -590,7 +590,7 @@ impl Init {
                     },
                 }),
                 definition_provider: Some(OneOf::Left(true)),
-                declaration_provider: Some(DeclarationCapability::Simple(true)),
+                references_provider: Some(OneOf::Left(true)),
                 completion_provider: Some(CompletionOptions {
                     trigger_characters: Some(vec![
                         String::from("#"),
